fix: Edit pinecone delete_index method

+++ app/services/pinecone_service.py
@@ -1,5 +1,7 @@
 import os
-from pinecone import Pinecone, ServerlessSpec
+
+from pinecone import Pinecone
+
 from app.services.openai_service import get_embedding
 
 # Initialize Pinecone with API key
@@ -12,15 +14,15 @@ pc = Pinecone(api_key=PINECONE_API_KEY)
 
 EMBEDDING_DIMENSION = 1536
 
-def embed_chunks_and_upload_to_pinecone(chunks, index_name):
 
+def embed_chunks_and_upload_to_pinecone(chunks, index_name):
     if index_name in pc.list_indexes():
         print("\nIndex already exists. Deleting index ...")
         pc.delete_index(name=index_name)
 
     print("\nCreating a new index: ", index_name)
     pc.create_index(name=index_name,
-                          dimension=EMBEDDING_DIMENSION, metric='cosine')
+                    dimension=EMBEDDING_DIMENSION, metric='cosine')
 
     index = pc.Index(index_name)
     # print(index.describe_index_stats())
@@ -38,6 +40,7 @@ def embed_chunks_and_upload_to_pinecone(chunks, index_name):
 
     print(f"\nUploaded {len(chunks)} chunks to Pinecone index\n'{index_name}'.")
 
+
 def get_most_similar_chunks_for_query(query, index_name):
     print("\nEmbedding query using OpenAI ...")
     question_embedding = get_embedding(query)
@@ -49,6 +52,11 @@ def get_most_similar_chunks_for_query(query, index_name):
 
     return context_chunks
 
+
 def delete_index(index_name):
-    pc.delete_index(name=index_name)
-    print(f"Index {index_name} deleted successfully")
+    if index_name in pc.list_indexes():
+        print("\nDeleting index ...")
+        pc.delete_index(name=index_name)
+        print(f"Index {index_name} deleted successfully")
+    else:
+        print("\nNo index to delete!")
